# docs/api-swagger/swagger.yaml
definitions:
...
  handlers.CreateRunRequest:
    properties:
      concurrency:
        description: Concurrency control
        type: integer
      docker_image:
        description: Docker image to use when runner_type=docker
        type: string
      empty_target:
        description: Run without target (generates placeholder target)
        type: boolean
      flow:
        description: Workflow identification
        type: string
      heuristics_check:
        description: 'Heuristics check level: none, basic, advanced'
        type: string
      module:
        description: Module workflow name
        type: string
      notify_on_complete:
        description: Send notification when run completes
        type: boolean
      params:
        additionalProperties:
          type: string
        type: object
      priority:
        description: Priority and timeout
        type: string
      repeat:
        description: Repeat run after completion
        type: boolean
      repeat_wait_time:
        description: Wait time between repeats (e.g., 30s, 20m, 10h, 1d)
        type: string
      run_mode:
        description: 'local, distributed, cloud (default: local)'
        type: string
      runner_type:
        description: Runner configuration
        type: string
      schedule:
        description: Scheduling options
        type: string
      schedule_enabled:
        description: Enable scheduled execution
        type: boolean
      skip_validation:
        description: Validation
        type: boolean
      ssh_host:
        description: SSH host when runner_type=ssh
        type: string
      target:
        type: string
      target_file:
        description: Path to file containing targets (one per line)
        type: string
      targets:
        description: Multi-target support
        items:
          type: string
        type: array
      threads_hold:
        description: Execution options (mirrors CLI flags)
        type: integer
      timeout:
        description: Timeout in minutes (0 = no timeout)
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /golish/api/runs:
    get:
      description: Get a paginated list of workflow runs with optional filters
      parameters:
      - default: 0
        description: Number of records to skip
        in: query
        name: offset
        type: integer
      - default: 20
        description: Maximum number of records to return
        in: query
        name: limit
        type: integer
      - description: Filter by status (pending, running, completed, failed, cancelled)
        in: query
        name: status
        type: string
      - description: Filter by workflow name
        in: query
        name: workflow
        type: string
      - description: Filter by target (partial match)
        in: query
        name: target
        type: string
      - description: Filter by workspace name (exact match)
        in: query
        name: workspace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of runs
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List runs
      tags:
      - Runs
    post:
      consumes:
      - application/json
      description: Execute a workflow against one or more targets. Supports multiple
        targets via array or file, concurrency control, priority levels, custom timeouts,
        runner configuration (host/docker/ssh), and scheduling via cron expressions.
      parameters:
      - description: Run configuration with optional priority, timeout, runner config,
          and scheduling
        in: body
        name: run
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateRunRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Run started
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Workflow not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new run
      tags:
      - Runs
  /golish/api/runs/{id}:
    delete:
      description: Cancel a running workflow execution. This will terminate all running
        processes associated with the run.
      parameters:
      - description: Run ID or RunID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Run cancelled
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Run cannot be cancelled
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Run not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cancel a run
      tags:
      - Runs
    get:
      description: Get details of a specific run by ID, including steps and artifacts
      parameters:
      - description: Run ID or RunID
        in: path
        name: id
        required: true
        type: string
      - default: false
        description: Include step results
        in: query
        name: include_steps
        type: boolean
      - default: false
        description: Include artifacts
        in: query
        name: include_artifacts
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Run details
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Run not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get run details
      tags:
      - Runs
...
  /golish/api/snapshot-download/{workspace_name}:
    get:
      description: Compress a workspace folder into a zip file and download it
      parameters:
      - description: Workspace name
        in: path
        name: workspace_name
        required: true
        type: string
      produces:
      - application/zip
      responses:
        "200":
          description: Zip file download
          schema:
            type: file
        "404":
          description: Workspace not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to create snapshot
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Download workspace snapshot
      tags:
      - Snapshots
...
  /golish/api/upload-file:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file containing a list of inputs (targets, URLs, etc.)
        for later use in runs
      parameters:
      - description: Input file to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: File uploaded with path
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Upload input file
      tags:
      - Files
...
  /golish/api/workflow-upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a raw YAML workflow file and save it to the workflows directory
      parameters:
      - description: Workflow YAML file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Workflow uploaded
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or YAML
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Upload workflow file
      tags:
      - Workflows
...
swagger: "2.0"